requirements.txt
fastapi>=0.68.1
uvicorn>=0.15.0
numpy>=1.21.2
torch>=1.9.0
transformers>=4.11.3
soundfile>=0.12.1
suno-bark
python-multipart>=0.0.5 